Add more components & improve the rendering system

+++ src/firefly/Frame.cxx
@@ -4,7 +4,7 @@ namespace firefly {
 
 Frame::Frame(const TexturePointer& texture, 
 	const SDL_Rect& sourceRect,
-	int durationMs): 
+	uint64_t durationMs): 
 	_texture(texture),
 	_sourceRect(sourceRect),
 	_durationMs(durationMs) {
@@ -21,7 +21,7 @@ SDL_Rect Frame::getSourceRect() const {
 	return _sourceRect;
 }
 
-int Frame::getDuration() const {
+uint64_t Frame::getDuration() const {
 	return _durationMs;
 }
